requirements.txt
# HTTP 클라이언트
httpx>=0.24,<0.26
aiohttp==3.9.1

# 웹 스크래핑
beautifulsoup4==4.12.3
# selectolax==0.3.21  # Python 3.13 호환 문제
# lxml==5.1.0  # 시스템 의존성 필요, html.parser 사용

# 데이터 모델
pydantic>=2.0
pydantic-settings>=2.0

# 데이터베이스
supabase==2.3.4

# AI/LLM
openai>=1.0
anthropic>=0.8
google-generativeai>=0.3

# 알림
slack-sdk>=3.0
discord-webhook>=1.0

# 이메일
resend>=0.7

# 유틸리티
python-dotenv>=1.0
python-dateutil>=2.8
rich>=13.0

# 테스트
pytest>=7.0
pytest-asyncio>=0.23
pytest-cov>=4.0
